backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
Pillow>=10.3.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.44.0
chromadb>=0.5.0
sentence-transformers>=3.0.1
requests>=2.32.0
